requirements.txt
selenium==4.4.3
google==3.0.0
python-dotenv==1.0.0
PyPDF2==3.0.1
google-genai